Migrate to maven central publishing
    * Update Gradle to 6.7.1 because it supports in memory signing
